move options to an internal package
