+++ flink/src/main/java/io/delta/flink/internal/options/BooleanOptionTypeConverter.java
@@ -1,4 +1,4 @@
-package io.delta.flink.options;
+package io.delta.flink.internal.options;
 
 /**
  * Implementation of {@link OptionTypeConverter} that validates values for
