+++ flink/src/main/java/io/delta/flink/internal/options/DefaultOptionTypeConverter.java
@@ -1,4 +1,4 @@
-package io.delta.flink.options;
+package io.delta.flink.internal.options;
 
 /**
  * An implementation of {@link OptionTypeConverter} interface to convert {@link DeltaConfigOption}
