+++ flink/src/main/java/io/delta/flink/internal/options/NonNegativeNumberTypeConverter.java
@@ -1,4 +1,4 @@
-package io.delta.flink.options;
+package io.delta.flink.internal.options;
 
 import org.apache.flink.util.StringUtils;
