+++ flink/src/main/java/io/delta/flink/internal/options/OptionTypeConverter.java
@@ -1,4 +1,4 @@
-package io.delta.flink.options;
+package io.delta.flink.internal.options;
 
 /**
  * Converter and validator for {@link DeltaConfigOption} values.
