+++ flink/src/main/java/io/delta/flink/internal/options/OptionValidator.java
@@ -1,4 +1,4 @@
-package io.delta.flink.options;
+package io.delta.flink.internal.options;
 
 import java.util.Map;
