+++ flink/src/main/java/io/delta/flink/internal/options/StartingVersionOptionTypeConverter.java
@@ -1,4 +1,4 @@
-package io.delta.flink.options;
+package io.delta.flink.internal.options;
 
 import java.util.regex.Pattern;
