+++ flink/src/main/java/io/delta/flink/source/internal/enumerator/SplitEnumeratorProvider.java
@@ -2,7 +2,7 @@ package io.delta.flink.source.internal.enumerator;
 
 import java.io.Serializable;
 
-import io.delta.flink.options.DeltaConnectorConfiguration;
+import io.delta.flink.internal.options.DeltaConnectorConfiguration;
 import io.delta.flink.source.internal.state.DeltaEnumeratorStateCheckpoint;
 import io.delta.flink.source.internal.state.DeltaSourceSplit;
 import org.apache.flink.api.connector.source.Boundedness;
