+++ flink/src/main/java/io/delta/flink/source/internal/enumerator/supplier/SnapshotSupplier.java
@@ -1,6 +1,6 @@
 package io.delta.flink.source.internal.enumerator.supplier;
 
-import io.delta.flink.options.DeltaConnectorConfiguration;
+import io.delta.flink.internal.options.DeltaConnectorConfiguration;
 import io.delta.flink.source.internal.utils.TransitiveOptional;
 
 import io.delta.standalone.DeltaLog;
