+++ flink/src/test/java/io/delta/flink/source/RowDataDeltaSourceBuilderTestBase.java
@@ -5,9 +5,9 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Stream;
 
-import io.delta.flink.options.DeltaConnectorConfiguration;
+import io.delta.flink.internal.options.DeltaConnectorConfiguration;
 import io.delta.flink.source.internal.DeltaSourceOptions;
-import io.delta.flink.options.DeltaConfigOption;
+import io.delta.flink.internal.options.DeltaConfigOption;
 import io.delta.flink.source.internal.builder.DeltaSourceBuilderBase;
 import io.delta.flink.source.internal.exceptions.DeltaSourceValidationException;
 import org.apache.hadoop.conf.Configuration;
