+++ flink/src/test/java/io/delta/flink/source/internal/DeltaSourceOptionsTest.java
@@ -6,7 +6,7 @@ import java.util.Set;
 import static java.lang.reflect.Modifier.isPublic;
 import static java.lang.reflect.Modifier.isStatic;
 
-import io.delta.flink.options.DeltaConfigOption;
+import io.delta.flink.internal.options.DeltaConfigOption;
 import org.apache.flink.configuration.ConfigOption;
 import org.junit.Test;
 import static org.hamcrest.MatcherAssert.assertThat;
