+++ flink/src/test/java/io/delta/flink/source/internal/enumerator/ContinuousDeltaSourceSplitEnumeratorCheckpointingTest.java
@@ -5,7 +5,7 @@ import java.util.Collections;
 import java.util.Iterator;
 import static java.util.Collections.singletonMap;
 
-import io.delta.flink.options.DeltaConnectorConfiguration;
+import io.delta.flink.internal.options.DeltaConnectorConfiguration;
 import io.delta.flink.source.internal.DeltaSourceOptions;
 import io.delta.flink.source.internal.file.AddFileEnumerator;
 import io.delta.flink.source.internal.state.DeltaEnumeratorStateCheckpoint;
