+++ flink/src/test/java/io/delta/flink/source/internal/utils/TestOptions.java
@@ -2,7 +2,7 @@ package io.delta.flink.source.internal.utils;
 
 import java.util.UUID;
 
-import io.delta.flink.options.DeltaConfigOption;
+import io.delta.flink.internal.options.DeltaConfigOption;
 import org.apache.flink.configuration.ConfigOptions;
 
 public class TestOptions {
